Upstream version 1.7.1

+++ tracking/include/pcl/tracking/impl/particle_filter.hpp
@@ -1,11 +1,11 @@
 #ifndef PCL_TRACKING_IMPL_PARTICLE_FILTER_H_
 #define PCL_TRACKING_IMPL_PARTICLE_FILTER_H_
 
-#include <boost/random.hpp>
-
 #include <pcl/common/common.h>
 #include <pcl/common/eigen.h>
 #include <pcl/common/transforms.h>
+#include <pcl/tracking/boost.h>
+#include <pcl/tracking/particle_filter.h>
 
 template <typename PointInT, typename StateT> bool
 pcl::tracking::ParticleFilterTracker<PointInT, StateT>::initCompute ()
@@ -135,7 +135,7 @@ pcl::tracking::ParticleFilterTracker<PointInT, StateT>::normalizeWeight ()
       {
         if (particles_->points[i].weight != 0.0)
         {
-          particles_->points[i].weight = normalizeParticleWeight (particles_->points[i].weight, w_min, w_max);
+          particles_->points[i].weight = static_cast<float> (normalizeParticleWeight (particles_->points[i].weight, w_min, w_max));
         }
       }
     }
